+++ lib/compony/model_fields/anchormodel.rb
@@ -25,7 +25,8 @@ module Compony
           collection:    self.class.collect(anchormodel_class.all),
           label_method:  :first,
           value_method:  :second,
-          selected:      selected_key,
+          selected:      selected_key, # if used in select
+          checked:      selected_key, # if used in radio buttons
           include_blank: anchormodel_attribute.optional
         }.merge(input_opts)
         return form.input name || @name, **opts