+++ lib/compony/model_fields/datetime.rb
@@ -1,8 +1,8 @@
 module Compony
   module ModelFields
     class Datetime < Base
-      def value_for(data, controller: nil, **_)
-        return transform_and_join(data.send(@name), controller:) { |el| el.nil? ? nil : I18n.l(el) }
+      def value_for(data, controller: nil, format: nil, **_)
+        return transform_and_join(data.send(@name), controller:) { |el| el.nil? ? nil : I18n.l(el, format:) }
       end
     end
   end