+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix spacing problem with invisible action buttons
 - Pass the controller as argument to action buttons
 - Support option `:label` in `value_for` of date and datetime, support superfluous options in association
+- Support argument list in form generator
 
 # 0.0.7